+++ src/pages/TagCreator/index.js
@@ -3,40 +3,13 @@ import React, { useState,useEffect,useRef } from 'react';
 import axios from 'axios';
 import Box from '@material-ui/core/Box'
 import Container from '@material-ui/core/Container';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
-import NavBreadcrumbs from '../../components/NavBreadcrumbs'
+import { HeaderSection } from './HeaderSection'
 import { AffiliationQuestionSection } from "./AffilationQuestionSection";
 import { OrgNameSection,OrganizationSelectorSection,OrgChange } from './Organization'
 import { ProjectRepositorySection,ProjectRepositoryInput } from './ProjectRepository'
 import { AddTopicTagSection,AddTagsQuestion,NewTags,CopyPasteTags,AddMoreTags,CurrentTopicTagSection } from './TopicTagSection'
-import { TitleSection } from '../../components'
-import { makeStyles } from '@material-ui/core/styles'
 
 
-const crumbs = [{ href: '/home', name: 'Home' }, { href: '/tag-generator', name: 'Tag Generator' }]
-
-const useStyles = makeStyles((theme) => ({
-
-}))
-
-const HeaderSection = () => {
-  const classes = useStyles()
-  return (
-    <Box className='boxBackground'>
-      <Container style={{ paddingBottom:'40px' }}>
-        <NavBreadcrumbs crumbs={crumbs} color="secondary" />
-        <Grid container>
-          <Grid item xs={12}><Typography variant='h4' fontFamily='Space Mono' color='textSecondary' className={classes.textStyle}>Civic Tech Index</Typography></Grid>
-          <TitleSection>Tag Generator</TitleSection>
-          <Grid item xs={12}>
-            <Typography variant='h6' color='textSecondary' style={{ fontWeight:'500' , textAlign:'center' }} >Join the Civic Tech Index by submitting your open-source project.<br /> This process takes less than one minute to complete.</Typography>
-          </Grid>
-        </Grid>
-      </Container>
-    </Box>
-  )
-}
 
 /**
  * By removing matched text, we allow leeway in entering repository URL,
